requirements-serve.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
albumentations>=1.3.0
